requirements.txt
discord.py
PyGithub
PyNaCl
pandas
async-timeout